+++ src/main/java/es/udc/fic/vvs/contenido/Contenido.java
@@ -5,14 +5,14 @@ import java.util.List;
 public interface Contenido {
 
 	public String obtenerTitulo();
-	
-	public int obtenerDuracion();
-	
+
+	public double obtenerDuracion();
+
 	public List<Contenido> obtenerListaReproduccion();
-	
+
 	public List<Contenido> buscar(String subcadena);
-	
+
 	public void agregar(Contenido contenido, Contenido predecesor);
-	
+
 	public void eliminar(Contenido contenido);
 }
